+++ packages/qwik/src/methods/insert.ts
@@ -1,6 +1,7 @@
 import type {
   FieldArrayPath,
   FieldArrayPathValue,
+  FieldPath,
   FieldValues,
   FormStore,
   Maybe,
@@ -67,6 +68,12 @@ export function insert<
         const filterName = (value: string) =>
           value.startsWith(`${name}.`) && getPathIndex(name, value) >= index;
 
+        // Create function to sort by path index
+        const sortPathIndex = (
+          pathA: FieldPath<TFieldValues> | FieldArrayPath<TFieldValues>,
+          pathB: FieldPath<TFieldValues> | FieldArrayPath<TFieldValues>
+        ) => getPathIndex(name, pathA) - getPathIndex(name, pathB);
+
         // Create function to get next index name
         const getNextIndexName = <T extends string>(
           fieldOrFieldArrayName: T,
@@ -80,7 +87,7 @@ export function insert<
         // Move fields that come after new item one index further
         getFieldNames(form)
           .filter(filterName)
-          .sort()
+          .sort(sortPathIndex)
           .reverse()
           .forEach((fieldName) => {
             setFieldState(
@@ -93,7 +100,7 @@ export function insert<
         // Move field arrays that come after new item one index further
         getFieldArrayNames(form)
           .filter(filterName)
-          .sort()
+          .sort(sortPathIndex)
           .reverse()
           .forEach((fieldArrayName) => {
             setFieldArrayState(