+++ packages/qwik/src/methods/move.ts
@@ -70,6 +70,12 @@ export function move<
         }
       };
 
+      // Create function to sort by path index
+      const sortPathIndex = (
+        pathA: FieldPath<TFieldValues> | FieldArrayPath<TFieldValues>,
+        pathB: FieldPath<TFieldValues> | FieldArrayPath<TFieldValues>
+      ) => getPathIndex(name, pathA) - getPathIndex(name, pathB);
+
       // Create function to get previous index name
       const getPrevIndexName = <T extends string>(
         fieldOrFieldArrayName: T,
@@ -90,10 +96,12 @@ export function move<
         ) as T;
 
       // Create list of all affected field and field array names
-      const fieldNames = getFieldNames(form).filter(filterName).sort();
+      const fieldNames = getFieldNames(form)
+        .filter(filterName)
+        .sort(sortPathIndex);
       const fieldArrayNames = getFieldArrayNames(form)
         .filter(filterName)
-        .sort();
+        .sort(sortPathIndex);
 
       // Reverse names if "from" index is greater than "to" index
       if (fromIndex > toIndex) {