+++ packages/qwik/src/methods/remove.ts
@@ -1,5 +1,6 @@
 import type {
   FieldArrayPath,
+  FieldPath,
   FieldValues,
   FormStore,
   ResponseData,
@@ -53,6 +54,12 @@ export function remove<
       const filterName = (value: string) =>
         value.startsWith(`${name}.`) && getPathIndex(name, value) > index;
 
+      // Create function to sort by path index
+      const sortPathIndex = (
+        pathA: FieldPath<TFieldValues> | FieldArrayPath<TFieldValues>,
+        pathB: FieldPath<TFieldValues> | FieldArrayPath<TFieldValues>
+      ) => getPathIndex(name, pathA) - getPathIndex(name, pathB);
+
       // Create function to get previous index name
       const getPrevIndexName = <T extends string>(
         fieldOrFieldArrayName: T,
@@ -66,7 +73,7 @@ export function remove<
       // Move state of each field after the removed index back by one index
       getFieldNames(form)
         .filter(filterName)
-        .sort()
+        .sort(sortPathIndex)
         .forEach((fieldName) => {
           setFieldState(
             form,
@@ -78,7 +85,7 @@ export function remove<
       // Move state of each field array after the removed index back by one index
       getFieldArrayNames(form)
         .filter(filterName)
-        .sort()
+        .sort(sortPathIndex)
         .forEach((fieldArrayName) => {
           setFieldArrayState(
             form,