+++ packages/solid/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Remove unused `InitialValues` utility type
+
 ## v0.12.2 (April 29, 2023)
 
 - Fix sorting bug in `insert`, `move` and `remove` method (issue #61)